Use compile-time checks to account for Windows & Unix differences.
