+++ src/System.Drawing.Common/src/System.Drawing.Common.csproj
@@ -32,7 +32,6 @@
     <Reference Include="System.Runtime" />
     <Reference Include="System.Runtime.Extensions" />
     <Reference Include="System.Runtime.InteropServices" />
-    <Reference Include="System.Runtime.InteropServices.RuntimeInformation" />
     <Reference Include="System.ComponentModel" />
     <Reference Include="System.ComponentModel.Primitives" />
     <Reference Include="System.Threading" />
@@ -50,7 +49,7 @@
     <Compile Include="System\Drawing\Pen.cs" />
     <Compile Include="System\Drawing\Pens.cs" />
     <Compile Include="System\Drawing\RotateFlipType.cs" />
-    <Compile Include="System\Drawing\SystemPens.cs" />    
+    <Compile Include="System\Drawing\SystemPens.cs" />
     <Compile Include="System\Drawing\Printing\PrinterUnit.cs" />
     <Compile Include="System\Drawing\Printing\PreviewPageInfo.cs" />
     <Compile Include="System\Drawing\Printing\PrintEventHandler.cs" />
@@ -203,7 +202,6 @@
     <Compile Include="System\Drawing\ContentAlignment.cs" />
     <Compile Include="System\Drawing\Drawing2D\CustomLineCap.Windows.cs" />
     <Compile Include="System\Drawing\Drawing2D\GraphicsPath.cs" />
-    <Compile Include="System\Drawing\Drawing2D\Matrix.cs" />
     <Compile Include="System\Drawing\Drawing2D\LinearGradientBrush.cs" />
     <Compile Include="System\Drawing\Drawing2D\PathGradientBrush.cs" />
     <Compile Include="System\Drawing\Font.cs" />
@@ -240,6 +238,7 @@
     <Compile Include="System\Drawing\Printing\PrintPageEvent.cs" />
     <Compile Include="System\Drawing\Printing\PrintPreviewGraphics.cs" />
     <Compile Include="System\Drawing\Printing\TriState.cs" />
+    <Compile Include="System\Drawing\Region.Windows.cs" />
     <Compile Include="System\Drawing\StringFormat.cs" />
     <Compile Include="System\Drawing\SystemColors.cs" />
     <Compile Include="System\Drawing\SystemFonts.cs" />
@@ -312,6 +311,7 @@
     <Compile Include="Unix\System.Drawing\Image.cs" />
     <Compile Include="Unix\System.Drawing\KnownColors.cs" />
     <Compile Include="System\Drawing\Pen.Unix.cs" />
+    <Compile Include="System\Drawing\Region.Unix.cs" />
     <Compile Include="Unix\System.Drawing\SRDescriptionAttribute.cs" />
     <Compile Include="Unix\System.Drawing\StringFormat.cs" />
     <Compile Include="Unix\System.Drawing\SystemColors.cs" />
@@ -320,10 +320,10 @@
     <Compile Include="Unix\System.Drawing\ToolboxBitmapAttribute.cs" />
     <Compile Include="System\Drawing\Drawing2D\AdjustableArrowCap.Unix.cs" />
     <Compile Include="System\Drawing\Drawing2D\CustomLineCap.Unix.cs" />
-    <Compile Include="Unix\System.Drawing.Drawing2D\GraphicsPath.cs" />\
-    <Compile Include="Unix\System.Drawing.Drawing2D\Matrix.cs" />\
+    <Compile Include="Unix\System.Drawing.Drawing2D\GraphicsPath.cs" />
+    <Compile Include="Unix\System.Drawing.Drawing2D\Matrix.cs" />
     <Compile Include="Unix\System.Drawing.Drawing2D\LinearGradientBrush.cs" />
-    <Compile Include="Unix\System.Drawing.Drawing2D\PathGradientBrush.cs" />\
+    <Compile Include="Unix\System.Drawing.Drawing2D\PathGradientBrush.cs" />
     <Compile Include="Unix\System.Drawing.Imaging\BitmapData.cs" />
     <Compile Include="Unix\System.Drawing.Imaging\ImageFormat.cs" />
     <Compile Include="Unix\System.Drawing.Imaging\Metafile.cs" />
